refactor extracted constants into separate files
