+++ src/shared/lib/constants/initialCurrentPage.ts
@@ -1,0 +1,1 @@
+export const initialCurrentPage = '1'
