+++ src/shared/lib/constants/selectOptionsForBan.ts
@@ -1,0 +1,5 @@
+export const selectOptionsForBan = [
+  {text: 'Not selected', value: 'ALL'},
+  { text: 'Blocked', value: 'BLOCKED' },
+  { text: 'Not blocked', value: 'UNBLOCKED' },
+]
