+++ src/shared/lib/constants/selectOptionsForPagination.ts
@@ -1,0 +1,6 @@
+export const selectOptionsForPagination = [
+  { text: '8', value: '8' },
+  { text: '15', value: '15' },
+  { text: '30', value: '30' },
+  { text: '50', value: '50' },
+]
