+++ src/shared/lib/constants/userTabsVariants.ts
@@ -1,0 +1,6 @@
+export const  userTabsVariants = {
+  uploadedPhotos: 'Uploaded photos',
+  payments: 'Payments',
+  followers: 'Followers',
+  following: 'Following'
+}
