+++ src/shared/lib/index.ts
@@ -1,3 +1,7 @@
 export {useCustomToast} from './hooks/useCustomToast'
 export {MAIN_DOMAIN} from './constants/globalConstants'
 export { formatDate, combineFirstLastName } from './helpers/transformData'
+export { userTabsVariants } from './constants/userTabsVariants'
+export { selectOptionsForPagination } from './constants/selectOptionsForPagination'
+export { selectOptionsForBan } from './constants/selectOptionsForBan'
+export { initialCurrentPage } from './constants/initialCurrentPage'
